Service2/requirements.txt
boto3==1.35.33
botocore==1.35.33
mysql_connector_repackaged==0.3.1
pika==1.3.2
python-dotenv==1.0.1
Requests==2.32.3
